requirements.txt
streamlit>=1.32
plotly>=5.19
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
hmmlearn>=0.3
duckdb>=0.10
fredapi>=0.5
requests>=2.31
python-dotenv>=1.0
